usr.sbin/btxld compile fixes

+++ usr.sbin/btxld/btxld.c
@@ -253,7 +253,7 @@ btxld(const char *iname)
 	}
 	btx.btx_pgctl -= cwr;
 	btx.btx_entry = Eflag ? centry : ihdr.entry;
-	if (snprintf(name, sizeof(name), "%s.tmp", oname) >= sizeof(name))
+	if (snprintf(name, sizeof(name), "%s.tmp", oname) >= (int)sizeof(name))
 		errx(2, "%s: Filename too long", oname);
 	if ((fdo = open(name, O_CREAT | O_TRUNC | O_WRONLY, 0666)) == -1)
 		err(2, "%s", name);
@@ -469,7 +469,7 @@ writex(int fd, const void *buf, size_t nbyte)
 
 	if ((n = write(fd, buf, nbyte)) == -1)
 		err(2, "%s", tname);
-	if (n != nbyte)
+	if (n != (ssize_t)nbyte)
 		errx(2, "%s: Short write", tname);
 }
